Remove Event component
